Extract lookup operator helpers
